pyproject.toml
[project]
name = "DiscordBot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Collin Barber", email = "collinjbarber@gmail.com"},
]
dependencies = [
    "pydantic-settings>=2.4.0",
    "pydantic>=2.9.0",
    "supabase>=2.7.4",
    "discord-py>=2.4.0",
    "discord-py-interactions>=5.13.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
